Updated poster analysis (renamed structure, docs for tmdb & bechdeltest)
