+++ docs/library/data/bechdeltestcom.md
@@ -1,0 +1,6 @@
+# Bechdeltestcom
+
+::: bechdelai.data.bechdeltestcom
+    options:
+      heading_level: 3
+
