+++ docs/library/image/face_detection.md
@@ -1,0 +1,5 @@
+# Face Detection
+
+::: bechdelai.image.face_detection
+    options:
+      heading_level: 3
