+++ docs/library/image/img.md
@@ -1,0 +1,5 @@
+# Img
+
+::: bechdelai.image.img
+    options:
+      heading_level: 3
