+++ docs/library/index.md
@@ -1,5 +1,6 @@
 
-# API documentation
+# Library documentation
 Find in this section all API information of the libary
 
-WIP
+!!! warning
+    Under construction
