+++ pyproject.toml
@@ -32,15 +32,11 @@ requests = ">=2.26.0"
 beautifulsoup4 = ">=4.10.0"
 chardet = ">=4.0.0"
 openpyxl = "^3.0.10"
-streamlit = "^1.9.2"
 
 [tool.poetry.dev-dependencies]
 pywin32 = { version = "225", markers = "sys_platform == 'win32'" }
 jupyter = "^1.0.0"
-mkdocs = "^1.2.1"
-mkdocs-material = "^7.1.10"
-mknotebooks = "^0.7.0"
-mkdocstrings = "^0.15.2"
+mkdocs = "^1.4.1"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
